+++ src/lib.rs
@@ -47,6 +47,7 @@ pub mod client;
 pub mod config;
 pub mod endpoint;
 pub mod error;
+pub mod limits;
 pub mod merchant;
 pub mod money;
 pub mod notify;